Pick date from ActionBar; Added DiningHall/Meal Fragments design
