+++ app/src/main/res/values/strings.xml
@@ -7,7 +7,7 @@
     <string name="title_dinner">Dinner</string>
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
-    <string name="action_example">Example action</string>
+    <string name="action_pick_date">Pick Date</string>
     <string name="action_settings">Settings</string>
     <string name="default_menu_text">Trying to fetch menu data</string>
     <string name="menu_date">Select a Date</string>
